requirements.txt
amqp==5.1.1
autopep8==2.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
greenlet==2.0.2
idna==3.4
kombu==5.2.4
marshmallow==3.19.0
packaging==23.0
prompt-toolkit==3.0.38
pycodestyle==2.10.0
python-dotenv==1.0.0
pytz==2022.7.1
requests==2.28.2
six==1.16.0
SQLAlchemy==2.0.5.post1
typing_extensions==4.5.0
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.6
